+++ src/cmd/mod.rs
@@ -6,7 +6,7 @@ pub mod pos;
 pub mod range;
 pub mod rsw;
 pub mod sliding;
-// pub mod stat;
+pub mod stat;
 pub mod status;
 pub mod tsv;
 pub mod wave;