+++ CHANGELOG.md
@@ -8,7 +8,9 @@
   * `garr env` creates sqls
   * Stats of ctg and rsw
 
-* `garr stat` - `await` makes the compilation extremely slow
+* `garr stat`
+  * `Polars` can't read sqls, so use the built-in queries
+  * `Datafusion` makes the compilation extremely slow
 
 ## 0.1.0 - 2021-08-27